Move: ApiResponseVoid.kt(remote.model.response.community -> remote.model.response)
