+++ core/remote/src/main/java/kr/co/remote/model/response/ApiResponseVoid.kt
@@ -1,4 +1,4 @@
-package kr.co.remote.model.response.community
+package kr.co.remote.model.response
 
 import kotlinx.serialization.Serializable
